moved tmux to a new package
