+++ tmux/TmuxSession_test.go
@@ -1,4 +1,4 @@
-package asciichgolangpublic
+package tmux
 
 import (
 	"testing"
